apps/server-upload/python-cluster/requirements.txt
cython>=0.29.24,<3.0
numpy==1.19.5
scikit-learn==1.0.2
pandas==1.4.0
flask==2.0.1
umap-learn==0.5.3
hdbscan==0.8.27
gunicorn==20.1.0
flask-cors==3.0.10 